Checkin the option of LOBPCGSolve2.
This option does not use the deflation to stabilize LOBPCG, but uses orthogonalization of the basis functions.

There are multiple ways to perform orthogonalization, and here eigenvalue solver is used.  This is not the most efficient nor the most stable way, but the stability should improve over LOBPCGSolve1.

TODO: Cleanup the LOBPCGSolve2 code.