Fix a bug in computing the force using PEXSI.
In CalculateForceDM, when computing the contribution from the nonlocal
pseudopotentials, the previous version did not check whether the
pseudopotential vector is zero or not in a given element.  This may
cause processors to request for matrix blocks in the density matrix
corresponding to a zero matrix block in the Hamiltonian matrix.

This problem is fixed and passed for a Li954 system.