New version of DG-PEXSI.
The data communication routine are data distribution of PEXSI is
updated.

The DistSparseMat <-> DistElemMat are updated to

DistSparseMatToDistElemMat3 and
DistElemMatToDistSparseMat3

These two routines assume that PEXSI only uses a subset of processors
with number less than the number of elements.  As a result, all
communication can be performed in local communicators.