Improve DistSparseMatToDistElemMat3 routine.
In the previous implementation, the inverse mapping of elemBasisIdx
needs to be constructed each time when this subroutine is called.  This
is quite time consuming for large matrices.

In the current implementation, the inverse mapping is saved in
HamiltonianDG class, and is constructed only once, together with
elemBasisIdx.  Correspondingly, basisInvIdx is added as an input
parameter in DistSparseMatToDistElemMat3.