Clean up some routines for force calculation in PWDFT.
Introduce coarse to fine grid mapping to make it easier for dual-grid
transformation.

Settle down for the formulation of computing the force without using any
derivatives of the local or nonlocal pseudopotential. This could make
code easier to handle in the future. The accuracy is fully comparable to
those using the derivatives in the dual grid implementation.