Implement the structure factor in PWDFT
o Use CalculateAtomDensity in periodtable and hamiltonian to generate
the origin-centered atomic density with a structure factor
implementation.

This implementation could be useful for future pseudocharge
improvements if necessary

o Tested for LiH dimer system in PWDFT.

o Replace reference to periodtable structures in the code by class
member functions. The next step is to support both HGH and ONCV
pseudopotentials