Temporary bug fix for the Fourier interpolation
This seems to have a MAJOR impact in terms of the convergence behavior.

The behavior of numGrid[d]/2+1 for when numGrid[d] is an odd number is fixed.

The behavior of numGrid[d]/2+1 when numGrid[d] is an even number seems much
more difficult to fix and will be addressed later.

esdf.cpp has a TEMPORARY fix to make sure all numGrid[d] is odd