Move the decision of XCFamily etc from esdf to Hamiltonian class.
Introduce SetupXC / DestroyXC in Hamiltonian to decide the XC
functional. At the next step, this will allow us to use the PBE
functional to initiate a hybrid calculation, in order to solve the
problem that HSE without exact exchange can lead to a very large number
of SCF steps for some systems.

The decision to invoke CalculateRho is also moved to Hamiltonian