Collapse Hamiltonian and KohnSham classes.
Remove the inheritance structure. It is unclear whether the inheritance
structure would benefit future developments (if at all). For future
functionalties that are very different from Gamma point calculations
(for example, k-points), a different folder should be used instead.

The current structure (with components in spinor and density) should
still be good enough to accomodate spinful calculations (collinear as
well as non-colinear)