Fix a problem in reading UPF files
When reading a UPF file (like a HGH file), sometimes an attribute starts
with a new line without a space in front of it. This causes
periodtable.cpp to complain. This is now fixed by changing `get_attr` in
utility.cpp. A better approach is to use a more proper regex search
method (but this can wait when it becomes necessary)

Although the current code can read HGH file now, it does not run yet
since HGH requires non-diagonal DIJ matrix. This requires diagonalizing
a small DIJ matrix and modify the calculation of the nonlocal
pseudopotentials (previously in the BIN file format, this was done at
the MATLAB level), and will be implemented next.